Move EZBoard to 1769
Move EZBoard to 1769 - This board uses the 1769
